% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_checks.R
\name{check_persistent_requirements}
\alias{check_persistent_requirements}
\title{Check persistent watcher requirements}
\usage{
check_persistent_requirements()
}
\value{
Logical indicating if persistent watcher requirements are met
}
\description{
Check persistent watcher requirements
}
\keyword{internal}
