% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_checks.R
\name{check_platform_requirements}
\alias{check_platform_requirements}
\title{Check platform-specific requirements}
\usage{
check_platform_requirements()
}
\description{
Check platform-specific requirements
}
\keyword{internal}
