% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_checks.R
\name{check_sqlite_requirements}
\alias{check_sqlite_requirements}
\title{Check SQLite availability and accessibility}
\usage{
check_sqlite_requirements()
}
\description{
Check SQLite availability and accessibility
}
\keyword{internal}
