% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{cleanup_inactive_processes}
\alias{cleanup_inactive_processes}
\title{Clean up inactive process records}
\usage{
cleanup_inactive_processes(db_path)
}
\arguments{
\item{db_path}{Path to watcher database}
}
\value{
Number of records cleaned up
}
\description{
Clean up inactive process records
}
\keyword{internal}
