% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{cleanup_old_databases}
\alias{cleanup_old_databases}
\title{Clean up old databases and orphaned resources}
\usage{
cleanup_old_databases(dir, max_age = 7)
}
\arguments{
\item{dir}{Directory to clean}

\item{max_age}{Maximum age in days before cleanup}
}
\description{
Clean up old databases and orphaned resources
}
\keyword{internal}
