% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{cleanup_old_files}
\alias{cleanup_old_files}
\title{Clean up old vigil files}
\usage{
cleanup_old_files(dir, max_age = 7)
}
\arguments{
\item{dir}{Directory to clean}

\item{max_age}{Maximum age in days before cleanup}
}
\description{
Clean up old vigil files
}
\keyword{internal}
