% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{cleanup_watcher_database}
\alias{cleanup_watcher_database}
\title{Clean up a watcher database}
\usage{
cleanup_watcher_database(db_path, force = FALSE)
}
\arguments{
\item{db_path}{Path to watcher database}

\item{force}{Whether to force cleanup of persistent watchers}
}
\value{
Logical indicating success
}
\description{
Clean up a watcher database
}
\keyword{internal}
