% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{cleanup_watcher_files}
\alias{cleanup_watcher_files}
\title{Clean up watcher files}
\usage{
cleanup_watcher_files(id, force = FALSE)
}
\arguments{
\item{id}{Watcher ID}

\item{force}{Whether to force cleanup of persistent watcher files}
}
\description{
Clean up watcher files
}
\keyword{internal}
