% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{cleanup_watcher_resources}
\alias{cleanup_watcher_resources}
\title{Clean up watcher resources}
\usage{
cleanup_watcher_resources(db_path, force = FALSE)
}
\arguments{
\item{db_path}{Path to watcher database}

\item{force}{Whether to force cleanup of persistent watcher}
}
\description{
Clean up watcher resources
}
\keyword{internal}
