% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_watching.R
\name{convert_regex_to_vbs}
\alias{convert_regex_to_vbs}
\title{Convert R regex to VBScript compatible pattern}
\usage{
convert_regex_to_vbs(pattern)
}
\arguments{
\item{pattern}{Character string containing R regular expression}
}
\value{
Character string containing VBScript compatible regular expression
}
\description{
Convert R regex to VBScript compatible pattern
}
\note{
VBScript has limited regex support. Some R regex features may be removed
or simplified.
}
\keyword{internal}
