% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_watching.R
\name{create_unix_watcher_script}
\alias{create_unix_watcher_script}
\title{Create Unix watcher shell script}
\usage{
create_unix_watcher_script(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
Character string containing shell script code
}
\description{
Create Unix watcher shell script
}
\keyword{internal}
