% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_watching.R
\name{create_windows_watcher_script}
\alias{create_windows_watcher_script}
\title{Create Windows watcher VBScript}
\usage{
create_windows_watcher_script(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
Character string containing VBScript code
}
\description{
Create Windows watcher VBScript
}
\keyword{internal}
