% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnload}
\alias{.onUnload}
\title{Package unload function}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{Library path}
}
\description{
Cleans up non-persistent watchers when package is unloaded.
}
\keyword{internal}
