% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_sql_string}
\alias{escape_sql_string}
\title{Safely escape string for SQL}
\usage{
escape_sql_string(str)
}
\arguments{
\item{str}{String to escape}
}
\value{
Escaped string safe for SQL
}
\description{
Safely escape string for SQL
}
\keyword{internal}
