% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_sql_timestamp}
\alias{format_sql_timestamp}
\title{Format timestamp for SQLite storage}
\usage{
format_sql_timestamp(time = Sys.time())
}
\arguments{
\item{time}{POSIXt time object (defaults to current time)}
}
\value{
Character string in SQLite timestamp format
}
\description{
Format timestamp for SQLite storage
}
\keyword{internal}
