% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_script_path}
\alias{get_script_path}
\title{Get package script path}
\usage{
get_script_path(script_name)
}
\arguments{
\item{script_name}{Name of script to find}
}
\value{
Absolute path to script
}
\description{
Get package script path
}
\keyword{internal}
