% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_watcher_events}
\alias{get_watcher_events}
\title{Get events for a watcher}
\usage{
get_watcher_events(db_path, include_output = FALSE)
}
\arguments{
\item{db_path}{Path to watcher database}

\item{include_output}{Include callback output in results}
}
\value{
Tibble containing events
}
\description{
Get events for a watcher
}
\keyword{internal}
