% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_watcher_process}
\alias{get_watcher_process}
\title{Get watcher process information}
\usage{
get_watcher_process(db_path)
}
\arguments{
\item{db_path}{Path to watcher database}
}
\value{
List containing process info (pid, state)
}
\description{
Get watcher process information
}
\keyword{internal}
