% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_watching.R
\name{glob2regexp}
\alias{glob2regexp}
\title{Convert glob pattern to extended regular expression}
\usage{
glob2regexp(pattern)
}
\arguments{
\item{pattern}{Glob pattern (e.g., "*.csv")}
}
\value{
Extended regular expression
}
\description{
Convert glob pattern to extended regular expression
}
\keyword{internal}
