% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{is_active_watcher_file}
\alias{is_active_watcher_file}
\title{Check if file belongs to active watcher}
\usage{
is_active_watcher_file(file)
}
\arguments{
\item{file}{File path to check}
}
\value{
Boolean indicating if file belongs to active watcher
}
\description{
Check if file belongs to active watcher
}
\keyword{internal}
