% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{is_watcher_active}
\alias{is_watcher_active}
\title{Check if watcher database exists and is active}
\usage{
is_watcher_active(id)
}
\arguments{
\item{id}{Watcher identifier}
}
\value{
Logical indicating if watcher is active
}
\description{
Check if watcher database exists and is active
}
\keyword{internal}
