% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{kill_process}
\alias{kill_process}
\title{Kill a vigil watcher process}
\usage{
kill_process(pid, id)
}
\arguments{
\item{pid}{Process ID to kill}

\item{id}{Watcher ID}
}
\value{
Boolean indicating success
}
\description{
Kill a vigil watcher process
}
\keyword{internal}
