% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{kill_unix_process}
\alias{kill_unix_process}
\title{Kill a Unix process}
\usage{
kill_unix_process(pid, timeout)
}
\arguments{
\item{pid}{Process ID to kill}

\item{timeout}{Timeout in seconds}
}
\value{
Boolean indicating success
}
\description{
Kill a Unix process
}
\keyword{internal}
