% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_watcher_databases}
\alias{list_watcher_databases}
\title{List all watcher databases}
\usage{
list_watcher_databases(include_stopped = FALSE)
}
\arguments{
\item{include_stopped}{Include stopped watchers}
}
\value{
Character vector of database paths
}
\description{
List all watcher databases
}
\keyword{internal}
