% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{mark_process_inactive}
\alias{mark_process_inactive}
\title{Mark process as inactive in database}
\usage{
mark_process_inactive(db_path, pid)
}
\arguments{
\item{db_path}{Path to watcher database}

\item{pid}{Process ID to mark as inactive}
}
\value{
Logical indicating success
}
\description{
Mark process as inactive in database
}
\keyword{internal}
