% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_callback_env}
\alias{prepare_callback_env}
\title{Prepare environment for callback execution}
\usage{
prepare_callback_env(db_path, event_id)
}
\arguments{
\item{db_path}{Database path}

\item{event_id}{Event ID}
}
\value{
Environment ready for callback
}
\description{
Prepare environment for callback execution
}
\keyword{internal}
