% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_watcher_config}
\alias{read_watcher_config}
\title{Read watcher configuration from database}
\usage{
read_watcher_config(db_path)
}
\arguments{
\item{db_path}{Path to watcher database}
}
\value{
List containing configuration
}
\description{
Read watcher configuration from database
}
\keyword{internal}
