% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{register_persistent}
\alias{register_persistent}
\title{Register a watcher as a persistent service}
\usage{
register_persistent(id)
}
\arguments{
\item{id}{Watcher identifier}
}
\value{
Logical indicating success
}
\description{
Register a watcher as a persistent service
}
\keyword{internal}
