% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{register_persistent_linux}
\alias{register_persistent_linux}
\title{Register a Linux persistent watcher}
\usage{
register_persistent_linux(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
TRUE if successful
}
\description{
Register a Linux persistent watcher
}
\keyword{internal}
