% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{register_persistent_macos}
\alias{register_persistent_macos}
\title{Register a macOS persistent watcher}
\usage{
register_persistent_macos(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
TRUE if successful
}
\description{
Register a macOS persistent watcher
}
\keyword{internal}
