% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{register_persistent_windows}
\alias{register_persistent_windows}
\title{Register a Windows persistent watcher}
\usage{
register_persistent_windows(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
TRUE if successful
}
\description{
Register a Windows persistent watcher
}
\keyword{internal}
