% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.R
\name{save_watcher_output}
\alias{save_watcher_output}
\title{Save watcher output to a file}
\usage{
save_watcher_output(db_path, output_name = NULL)
}
\arguments{
\item{db_path}{Path to watcher database}

\item{output_name}{Optional name for output file}
}
\value{
Path to saved output file
}
\description{
Save watcher output to a file
}
\keyword{internal}
