% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_watching.R
\name{start_watcher}
\alias{start_watcher}
\title{Start a file watcher process}
\usage{
start_watcher(config)
}
\arguments{
\item{config}{List containing watcher configuration}
}
\value{
Process ID or database path depending on configuration
}
\description{
Start a file watcher process
}
\keyword{internal}
