% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{store_callback_info}
\alias{store_callback_info}
\title{Store callback information in database}
\usage{
store_callback_info(db_path, type, value)
}
\description{
Store callback information in database
}
\keyword{internal}
