% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{unregister_persistent}
\alias{unregister_persistent}
\title{Unregister a persistent watcher service}
\usage{
unregister_persistent(id)
}
\arguments{
\item{id}{Watcher identifier}
}
\value{
Logical indicating success
}
\description{
Unregister a persistent watcher service
}
\keyword{internal}
