% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{unregister_persistent_linux}
\alias{unregister_persistent_linux}
\title{Unregister a Linux persistent watcher}
\usage{
unregister_persistent_linux(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
TRUE if successful
}
\description{
Unregister a Linux persistent watcher
}
\keyword{internal}
