% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{unregister_persistent_macos}
\alias{unregister_persistent_macos}
\title{Unregister a macOS persistent watcher}
\usage{
unregister_persistent_macos(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
TRUE if successful
}
\description{
Unregister a macOS persistent watcher
}
\keyword{internal}
