% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{unregister_persistent_windows}
\alias{unregister_persistent_windows}
\title{Unregister a Windows persistent watcher}
\usage{
unregister_persistent_windows(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
TRUE if successful
}
\description{
Unregister a Windows persistent watcher
}
\keyword{internal}
