% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_callback}
\alias{validate_callback}
\title{Validate callback specifications}
\usage{
validate_callback(callback, db_path = NULL)
}
\arguments{
\item{callback}{Function, expression, or path to R script}

\item{db_path}{Path to watcher database for storing callback info}
}
\value{
List containing validated callback info
}
\description{
Validate callback specifications
}
\keyword{internal}
