% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{verify_persistent}
\alias{verify_persistent}
\title{Verify if a watcher is running as a persistent service}
\usage{
verify_persistent(id)
}
\arguments{
\item{id}{Watcher identifier}
}
\value{
Logical indicating if service is active
}
\description{
Verify if a watcher is running as a persistent service
}
\keyword{internal}
