% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{verify_persistent_linux}
\alias{verify_persistent_linux}
\title{Verify if a Linux persistent watcher is active}
\usage{
verify_persistent_linux(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
Boolean indicating if watcher is active
}
\description{
Verify if a Linux persistent watcher is active
}
\keyword{internal}
