% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.R
\name{verify_process_alive}
\alias{verify_process_alive}
\title{Check if a process is alive}
\usage{
verify_process_alive(pid)
}
\arguments{
\item{pid}{Process ID to check}
}
\value{
Logical indicating if process is running
}
\description{
Check if a process is alive
}
\keyword{internal}
