% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{verify_vigil_process}
\alias{verify_vigil_process}
\title{Verify if a process is a vigil watcher}
\usage{
verify_vigil_process(pid, id)
}
\arguments{
\item{pid}{Process ID to check}

\item{id}{Watcher ID}
}
\value{
Boolean indicating if process is a valid vigil watcher
}
\description{
Verify if a process is a vigil watcher
}
\keyword{internal}
