% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{verify_watcher_process}
\alias{verify_watcher_process}
\title{Verify if a process is a valid vigil watcher}
\usage{
verify_watcher_process(db_path)
}
\arguments{
\item{db_path}{Path to watcher database}
}
\value{
Boolean indicating if watcher is valid and running
}
\description{
Verify if a process is a valid vigil watcher
}
\keyword{internal}
