% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_management.R
\name{verify_windows_process}
\alias{verify_windows_process}
\title{Verify Windows process is a valid watcher}
\usage{
verify_windows_process(pid, id)
}
\arguments{
\item{pid}{Process ID to verify}

\item{id}{Watcher ID to check for}
}
\value{
Logical indicating if process is valid watcher
}
\description{
Verify Windows process is a valid watcher
}
\keyword{internal}
