% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_handling.R
\name{write_results}
\alias{write_results}
\title{Write captured results to database}
\usage{
write_results(con, event_id, handler)
}
\description{
Write captured results to database
}
\keyword{internal}
