% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_watching.R
\name{create_unix_watcher_script}
\alias{create_unix_watcher_script}
\title{Create Unix watcher shell script}
\usage{
create_unix_watcher_script(config)
}
\arguments{
\item{config}{Watcher configuration}
}
\value{
Character string containing shell script code
}
\description{
Create Unix watcher shell script
}
\details{
The pattern matching uses Extended Regular Expressions (ERE) which is compatible
with R's regex syntax. The pattern is passed to either inotifywait or fswatch
using their respective --regexp flags.

For inotifywait (Linux), see: https://linux.die.net/man/1/inotifywait
For fswatch (macOS), see: https://emcrisostomo.github.io/fswatch/doc/
}
\keyword{internal}
