% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{kill_all_watchers}
\alias{kill_all_watchers}
\title{Kill all file watchers}
\usage{
kill_all_watchers()
}
\description{
Stops all active file watchers and cleans up their resources.
}
\examples{
\dontrun{
# Kill all watchers
kill_all_watchers()
}
}
