% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{kill_watcher}
\alias{kill_watcher}
\title{Kill a file watcher}
\usage{
kill_watcher(id, timeout = 5)
}
\arguments{
\item{id}{Watcher ID to kill}

\item{timeout}{Timeout in seconds for kill operation}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Kill a file watcher
}
