% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_callback}
\alias{validate_callback}
\title{Validate Callback}
\usage{
validate_callback(callback)
}
\arguments{
\item{callback}{Function, expression, package function name, or path to R script}
}
\value{
List containing:
\describe{
\item{type}{One of "none", "package", "script", or "expression"}
\item{value}{The validated callback value}
}
}
\description{
Validates and normalizes callback specifications for file watchers.
}
\examples{
# Function callback
validate_callback(function(x) print(x))

# Package function
validate_callback("dplyr::select")

# R script
validate_callback("~/scripts/process_file.R")
}
\keyword{internal}
